move description tests over
