+++ src/server/pfs/server/testing/BUILD.bazel
@@ -76,6 +76,7 @@ go_test(
         "//src/pfs",
         "//src/pps",
         "//src/server/pfs",
+        "//src/server/pfs/pretty",
         "//src/server/pfs/server",
         "@com_github_docker_go_units//:go-units",
         "@com_github_google_go_cmp//cmp",
